+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Code/Domain Reload view, showing any issues raised by the Domain Reload Roslyn analyzer
 * Domain Reload diagnostic issues in the Settings view if Domain Reload is enabled
 * Diagnostics for unsupported APIs on WebGL target
+* Diagnostic for Static Batching being enabled whilst Entities Graphics is installed
 
 ### Changed
 * Improved issue description and suggestion strings, to make them clearer and more consistent